+++ src/main.cpp
@@ -14,12 +14,10 @@
 
 #define ALGINE_LOGGING
 
-#include "utils/shader_utils.cpp"
-#include "utils/texture_utils.cpp"
 #include "model.cpp"
 #include "lamp.cpp"
-#include "advanced_rendering.cpp"
-#include "utils/math.cpp"
+#include "algine.cpp"
+#include "sconstants.h"
 
 #include "lib/libcmf/loader.cpp"
 
@@ -27,7 +25,6 @@
 #include <GLFW/glfw3.h>
 
 #include <glm/glm.hpp>
-#include <glm/gtc/matrix_transform.hpp>
 
 #define SHADOW_MAP_RESOLUTION 1024
 
@@ -63,16 +60,6 @@ GLFWwindow* window;
 
 GLuint vao;
 
-// shaders
-const char *vertexShader = "src/resources/shaders/vertex_shader.glsl";
-const char *fragmentShader = "src/resources/shaders/fragment_shader.glsl";
-const char *vertexShadowShader = "src/resources/shaders/vertex_shadow_shader.glsl";
-const char *fragmentShadowShader = "src/resources/shaders/fragment_shadow_shader.glsl";
-const char *geometryShadowShader = "src/resources/shaders/geometry_shadow_shader.glsl";
-
-// shaders programs
-ShaderProgram shaderProgram, shaderShadowProgram;
-
 // matrices
 glm::mat4 projectionMatrix, viewMatrix, modelMatrix;
 
@@ -92,15 +79,22 @@ LampModel lightSources[LAMPS_COUNT];
 // shadow maps
 TextureArray shadowMaps;
 
-// ambient, diffuse, specular, normal mapping
-Mapping mapping;
+// renderer
+AlgineRenderer *renderer;
 
-// DOF variables
-float 
+float
+    // DOF variables
     focalDepth = 8.0f, 
     focalRange = 20.0f,
     dof_max_sigma = 6.0f,
-    dof_min_sigma = 0.0001f;
+    dof_min_sigma = 0.0001f,
+    
+    // diskRadius variables
+    diskRadius_k = 1.0f / 25.0f,
+    diskRadius_min = 0.0f,
+    
+    // other shadow variables
+    shadow_bias = 0.3f;
 
 void createProjectionMatrix() {
     projectionMatrix = glm::perspective(glm::radians(90.0f), (float) WIN_W / (float) WIN_H, 1.0f, 32.0f);
@@ -116,8 +110,8 @@ void createViewMatrix() {
 void window_size_callback(GLFWwindow* window, int width, int height) {
     WIN_W = width;
     WIN_H = height;
-    advren::SCR_W = WIN_W;
-    advren::SCR_H = WIN_H;
+    ALGINE_SCR_W = WIN_W;
+    ALGINE_SCR_H = WIN_H;
     createProjectionMatrix();
 }
 
@@ -135,7 +129,7 @@ void createLamp(LampModel &result, const glm::vec3 &pos, const glm::vec3 &color,
     result.kq = 0.0075f;
     result.shininess = 32;
 
-    result.loadLocations(shaderProgram.programId, id);
+    result.loadLocations(ALGINE_PROGRAM_ID_CS, id);
     result.initShadowMapping(SHADOW_MAP_RESOLUTION);
 }
 
@@ -220,27 +214,37 @@ void initGL() {
  * Loading and compiling shaders
  */
 void initShaders() {
-    shaderProgram.init(vertexShader, fragmentShader);
-    shaderProgram.loadValueId("a_Position", GLSL_ATTRIBUTE);
-	shaderProgram.loadValueId("a_Normal", GLSL_ATTRIBUTE);
-	shaderProgram.loadValueId("a_Texture", GLSL_ATTRIBUTE);
-	shaderProgram.loadValueId("a_Tangent", GLSL_ATTRIBUTE);
-	shaderProgram.loadValueId("a_Bitangent", GLSL_ATTRIBUTE);
-	shaderProgram.loadValueId("u_LampsCount", GLSL_UNIFORM);
-	shaderProgram.loadValueId("u_ViewPos", GLSL_UNIFORM);
-	shaderProgram.loadValueId("shadowMaps", GLSL_UNIFORM);
-	shaderProgram.loadValueId("u_Model", GLSL_UNIFORM);
-	shaderProgram.loadValueId("u_View", GLSL_UNIFORM);
-	shaderProgram.loadValueId("u_Projection", GLSL_UNIFORM);
-	shaderProgram.loadValueId("far_plane", GLSL_UNIFORM);
-    shaderProgram.loadValueId("focalDepth", GLSL_UNIFORM);
-    shaderProgram.loadValueId("focalRange", GLSL_UNIFORM);
+    ALGINE_PATHS_TO_SHADERS apts {
+        (const char*) "src/resources/shaders/vertex_shader.glsl",
+        (const char*) "src/resources/shaders/fragment_shader.glsl",
 
-    shaderShadowProgram.init(vertexShadowShader, fragmentShadowShader, geometryShadowShader);
-    shaderShadowProgram.loadValueId("a_Position", GLSL_ATTRIBUTE);
-	shaderShadowProgram.loadValueId("u_ModelMatrix", GLSL_UNIFORM);
-	shaderShadowProgram.loadValueId("lightPos", GLSL_UNIFORM);
-	shaderShadowProgram.loadValueId("far_plane", GLSL_UNIFORM);
+        (const char*) "src/resources/shaders/vertex_shadow_shader.glsl",
+        (const char*) "src/resources/shaders/geometry_shadow_shader.glsl",
+        (const char*) "src/resources/shaders/fragment_shadow_shader.glsl",
+
+        (const char*) "src/resources/shaders/vertex_blur_shader.glsl",
+        (const char*) "src/resources/shaders/fragment_blur_shader.glsl",
+
+        (const char*) "src/resources/shaders/vertex_blend_shader.glsl",
+        (const char*) "src/resources/shaders/fragment_blend_shader.glsl"
+    };
+
+    std::cout << "Compiling algine shaders\n";
+
+    //ALGINE_TEXTURE_MAPPING_MODE = ALGINE_TEXTURE_MAPPING_MODE_DISABLED;
+    //ALGINE_NORMAL_MAPPING_MODE = ALGINE_NORMAL_MAPPING_MODE_DISABLED;
+    //ALGINE_LIGHTING_MODE = ALGINE_LIGHTING_MODE_DISABLED;
+    //ALGINE_ATTENUATION_MODE = ALGINE_ATTENUATION_MODE_DISABLED;
+    //ALGINE_DOF_MODE = ALGINE_DOF_MODE_DISABLED;
+    //ALGINE_BLOOM_MODE = ALGINE_BLOOM_MODE_DISABLED;
+    //ALGINE_SHADOW_MAPPING_MODE = ALGINE_SHADOW_MAPPING_MODE_DISABLED;
+    //ALGINE_KERNEL_BLOOM_SIZE = 4;
+    ALGINE_KERNEL_DOF_SIZE = 3;
+    algine_compile_shaders(algine_create_shaders_sources(apts));
+    algine_load_locations();
+
+    renderer = new AlgineRenderer();
+    renderer->prepare();
 }
 
 /**
@@ -266,20 +270,20 @@ void init_ModelBuffers_ModelMapping() {
     // Lucy
     create_ModelBuffers_ModelMapping(
         "src/resources/models/LucyAngel/Stanfords_Lucy_Angel_triangulated.cmf",
-        "src/resources/textures/LucyAngel/Stanfords_Lucy_Angel_diffuse.jpg",
-        "src/resources/textures/LucyAngel/Stanfords_Lucy_Angel_diffuse.jpg",
-        "src/resources/textures/LucyAngel/Stanfords_Lucy_Angel_specular.jpg",
-        "src/resources/textures/LucyAngel/Stanfords_Lucy_Angel_normal.jpg",
+        "src/resources/textures/LucyAngel/Stanfords_Lucy_Angel_diffuse_1k.jpg",
+        "src/resources/textures/LucyAngel/Stanfords_Lucy_Angel_diffuse_1k.jpg",
+        "src/resources/textures/LucyAngel/Stanfords_Lucy_Angel_specular_1k.jpg",
+        "src/resources/textures/LucyAngel/Stanfords_Lucy_Angel_normal_1k.jpg",
         0
     );
 
     // Brickwall
     create_ModelBuffers_ModelMapping(
         "src/resources/models/brickwall/brickwall_triangulated_big.cmf",
-        "src/resources/textures/brickwall/brickwall_big.png",
-        "src/resources/textures/brickwall/brickwall_big.png",
-        "src/resources/textures/brickwall/brickwall_big.png",
-        "src/resources/textures/brickwall/brickwall_big_normal.png",
+        "src/resources/textures/brickwall/brickwall_big_2k.png",
+        "src/resources/textures/brickwall/brickwall_big_2k.png",
+        "src/resources/textures/brickwall/brickwall_big_2k.png",
+        "src/resources/textures/brickwall/brickwall_big_normal_2k.png",
         1
     );
 
@@ -336,8 +340,8 @@ void createModels() {
  * Creating light sources
  */
 void initLamps() {
-    Lamp::lightPosLocation = shaderShadowProgram.getValueId("lightPos");
-    Lamp::initShadowMatrices(shaderShadowProgram.programId);
+    Lamp::lightPosLocation = ALGINE_SS_LAMP_POSITION;
+    Lamp::initShadowMatrices(ALGINE_PROGRAM_ID_SS);
 
     createLamp(lightSources[0], glm::vec3(8.0f, 8.0f, -2.0f), glm::vec3(1.0f, 1.0f, 1.0f), 0);
     lamps[0].buffers = &mbuffers[3];
@@ -357,21 +361,32 @@ void initLamps() {
  */
 void initShadowMaps() {
     shadowMaps = TextureArray(10, LAMPS_COUNT);
-    shadowMaps.init(shaderProgram.programId, "shadowMaps");
+    shadowMaps.init(ALGINE_PROGRAM_ID_CS, ALGINE_NAME_CS_SHADOW_MAPS);
     for (size_t i = 0; i < LAMPS_COUNT; i++) shadowMaps.addTexture(i, lightSources[i].depthCubemap);
-    shadowMaps.bind(shaderProgram.programId);
+    shadowMaps.bind(ALGINE_PROGRAM_ID_CS);
+}
+
+void initShadowCalculation() {
+    glUseProgram(ALGINE_PROGRAM_ID_CS);
+    glUniform1f(ALGINE_CS_SHADOW_DISKRADIUS_K, diskRadius_k);
+    glUniform1f(ALGINE_CS_SHADOW_DISKRADIUS_MIN, diskRadius_min);
+    glUniform1f(ALGINE_CS_SHADOW_BIAS, shadow_bias);
+    glUseProgram(0);
 }
 
 /**
  * Initialize Depth of field
  */
 void initDOF() {
-    glUseProgram(advren::shaderBlurProgram.programId);
-    glUniform1f(advren::shaderBlurProgram.getValueId("max_sigma"), dof_max_sigma);
-    glUniform1f(advren::shaderBlurProgram.getValueId("min_sigma"), dof_min_sigma);
-    glUseProgram(shaderProgram.programId);
-    glUniform1f(shaderProgram.getValueId("focalDepth"), focalDepth);
-    glUniform1f(shaderProgram.getValueId("focalRange"), focalRange);
+    glUseProgram(ALGINE_PROGRAM_ID_BLUS_HORIZONTAL);
+    glUniform1f(ALGINE_BLUS_H_SIGMA_MAX, dof_max_sigma);
+    glUniform1f(ALGINE_BLUS_H_SIGMA_MIN, dof_min_sigma);
+    glUseProgram(ALGINE_PROGRAM_ID_BLUS_VERTICAL);
+    glUniform1f(ALGINE_BLUS_V_SIGMA_MAX, dof_max_sigma);
+    glUniform1f(ALGINE_BLUS_V_SIGMA_MIN, dof_min_sigma);
+    glUseProgram(ALGINE_PROGRAM_ID_CS);
+    glUniform1f(ALGINE_CS_FOCAL_DEPTH, focalDepth);
+    glUniform1f(ALGINE_CS_FOCAL_RANGE, focalRange);
     glUseProgram(0);
 }
 
@@ -382,19 +397,36 @@ void initDOF() {
  */
 void recycleAll() {
     glDeleteVertexArrays(1, &vao);
-    advren::recycle();
     for (size_t i = 0; i < MODEL_BUFFERS_MAPPING_COUNT; i++) {
         mbuffers[i].recycle();
         mmapping[i].recycle();
     }
+    delete renderer;
 }
 
 void sendLampsData() {
-	glUniform1i(shaderProgram.getValueId("u_LampsCount"), LAMPS_COUNT); // lamps count
-	su::setVec3(shaderProgram.getValueId("u_ViewPos"), cameraPos); // current camera position
-	glUniform1f(shaderProgram.getValueId("far_plane"), LAMP_FAR_PLANE); // far plane in shadow matrices
+	glUniform1i(ALGINE_CS_LAMPS_COUNT, LAMPS_COUNT); // lamps count
+	su::setVec3(ALGINE_CS_VIEW_POSITION, cameraPos); // current camera position
+	glUniform1f(ALGINE_CS_SHADOW_FAR_PLANE, LAMP_FAR_PLANE); // far plane in shadow matrices
 	for (size_t i = 0; i < LAMPS_COUNT; i++) lightSources[i].pushAll();
 }
+
+/* --- matrices --- */
+glm::mat4 getPVMMatrix() {
+    return projectionMatrix * viewMatrix * modelMatrix;
+}
+
+glm::mat4 getVMMatrix() {
+    return viewMatrix * modelMatrix;
+}
+
+void updateMatrices() {
+    su::setMat4(ALGINE_CS_MAT_PVM, getPVMMatrix());
+    su::setMat4(ALGINE_CS_MAT_VM, getVMMatrix());
+    su::setMat4(ALGINE_CS_MAT_MODEL, modelMatrix);
+    su::setMat4(ALGINE_CS_MAT_VIEW, viewMatrix);
+}
+/* --- --- */
 
 #define polygons buffers->polygons
 
@@ -403,14 +435,14 @@ void sendLampsData() {
  * @param shaderProgram
  * @param model
  */
-void drawModelDM(ShaderProgram &shaderProgram, Model &model) {
-	su::pointer(shaderProgram.getValueId("a_Position"), 3, model.getVerticesBuffer());
+void drawModelDM(Model &model) {
+	su::pointer(ALGINE_SS_IN_POSITION, 3, model.getVerticesBuffer());
 	modelMatrix = glm::mat4();
     modelMatrix = glm::translate(modelMatrix, model.origin);
     modelMatrix = glm::rotate(modelMatrix, model.angles[0], glm::vec3(1, 0, 0));
 	modelMatrix = glm::rotate(modelMatrix, model.angles[1], glm::vec3(0, 1, 0));
     modelMatrix = glm::rotate(modelMatrix, model.angles[2], glm::vec3(0, 0, 1));
-	su::setMat4(shaderProgram.getValueId("u_ModelMatrix"), modelMatrix);
+	su::setMat4(ALGINE_SS_MAT_MODEL, modelMatrix);
 	for (size_t i = 0; i < model.polygons.size; i++) glDrawArrays(model.polygons.array[i].type, model.polygons.array[i].start, model.polygons.array[i].count); // drawing
 }
 
@@ -425,44 +457,54 @@ void drawModelDM(ShaderProgram &shaderProgram, Model &model) {
  * @param shaderProgram
  * @param modelsArray
  */
-void drawModel(ShaderProgram &shaderProgram, Model &model) {
-	mapping.bindADSNTextures(model.ambientTexture, model.diffuseTexture, model.specularTexture, model.normalTexture);
-	su::pointer(shaderProgram.getValueId("a_Position"), 3, model.getVerticesBuffer());
-	su::pointer(shaderProgram.getValueId("a_Normal"), 3, model.getNormalsBuffer());
-	su::pointer(shaderProgram.getValueId("a_Tangent"), 3, model.getTangentsBuffer());
-	su::pointer(shaderProgram.getValueId("a_Bitangent"), 3, model.getBitangentsBuffer());
-	su::pointer(shaderProgram.getValueId("a_Texture"), 2, model.getTexCoordsBuffer());
+void drawModel(Model &model) {
+	bindADSNTextures(
+        ALGINE_CS_SAMPLER_MAPPING_AMBIENT, ALGINE_CS_SAMPLER_MAPPING_DIFFUSE, ALGINE_CS_SAMPLER_MAPPING_SPECULAR, ALGINE_CS_SAMPLER_MAPPING_NORMAL,
+        model.ambientTexture, model.diffuseTexture, model.specularTexture, model.normalTexture
+    );
+    
+	su::pointer(ALGINE_CS_IN_POSITION, 3, model.getVerticesBuffer());
+	su::pointer(ALGINE_CS_IN_NORMAL, 3, model.getNormalsBuffer());
+	su::pointer(ALGINE_CS_IN_TANGENT, 3, model.getTangentsBuffer());
+	su::pointer(ALGINE_CS_IN_BITANGENT, 3, model.getBitangentsBuffer());
+	su::pointer(ALGINE_CS_IN_TEXCOORD, 2, model.getTexCoordsBuffer());
 	modelMatrix = glm::mat4();
 	modelMatrix = glm::translate(modelMatrix, model.origin);
     modelMatrix = glm::rotate(modelMatrix, model.angles[0], glm::vec3(1, 0, 0));
 	modelMatrix = glm::rotate(modelMatrix, model.angles[1], glm::vec3(0, 1, 0));
     modelMatrix = glm::rotate(modelMatrix, model.angles[2], glm::vec3(0, 0, 1));
-	su::setMat4(shaderProgram.getValueId("u_Model"), modelMatrix);
+	updateMatrices();
 	for (size_t i = 0; i < model.polygons.size; i++) glDrawArrays(model.polygons.array[i].type, model.polygons.array[i].start, model.polygons.array[i].count); // drawing
 }
+
+#undef ambientTexture
+#undef diffuseTexture
+#undef specularTexture
+#undef normalTexture
+#undef polygons
 
 /**
  * Renders to depth cubemap
  */
-void renderToDepthMap(ShaderProgram &shaderProgram, GLuint index) {
+void renderToDepthMap(GLuint index) {
 	lightSources[index].begin();
 	lightSources[index].setLightPos();
 	lightSources[index].updateMatrices();
 	lightSources[index].setShadowMatrices();
 	glClear(GL_DEPTH_BUFFER_BIT);
 		
-	glEnableVertexAttribArray(shaderProgram.getValueId("a_Position"));
+	glEnableVertexAttribArray(ALGINE_SS_IN_POSITION);
 		
 	// drawing models
-	for (size_t i = 0; i < MODELS_COUNT; i++) drawModelDM(shaderProgram, models[i]);
+	for (size_t i = 0; i < MODELS_COUNT; i++) drawModelDM(models[i]);
 		
 	// drawing lamps
 	for (GLuint i = 0; i < LAMPS_COUNT; i++) {
 		if (i == index) continue;
-		drawModelDM(shaderProgram, lamps[i]);
+		drawModelDM(lamps[i]);
 	}
 		
-	glDisableVertexAttribArray(shaderProgram.getValueId("a_Position"));
+	glDisableVertexAttribArray(ALGINE_SS_IN_POSITION);
 	
 	lightSources[index].end();
 }
@@ -470,52 +512,56 @@ void renderToDepthMap(ShaderProgram &shaderProgram, GLuint index) {
 /**
  * Color rendering
  */
-void render(ShaderProgram &shaderProgram) {
-    // bloom
-	advren::begin();
+
+#define begin (renderer->*(renderer->begin_rendering))()
+#define end (renderer->*(renderer->end_rendering))()
+
+void render() {
+    // bloom, dof, blending
+    begin;
 
 	// view port to window size
 	glViewport(0, 0, WIN_W, WIN_H);
 	// updating view matrix (because camera position was changed)
 	createViewMatrix();
-	// sending updated matrices to vertex shader
-	su::setMat4(shaderProgram.getValueId("u_View"), viewMatrix);
-	su::setMat4(shaderProgram.getValueId("u_Projection"), projectionMatrix);
 	// sending lamps parameters to fragment shader
 	sendLampsData();
 	
-    glEnableVertexAttribArray(shaderProgram.getValueId("a_Position"));
-    glEnableVertexAttribArray(shaderProgram.getValueId("a_Normal"));
-    glEnableVertexAttribArray(shaderProgram.getValueId("a_Texture"));
+    glEnableVertexAttribArray(ALGINE_CS_IN_POSITION);
+    glEnableVertexAttribArray(ALGINE_CS_IN_NORMAL);
+    glEnableVertexAttribArray(ALGINE_CS_IN_TEXCOORD);
 	
 	// drawing
-	mapping.setNormalMappingEnabled(1); // with mapping
-    glEnableVertexAttribArray(shaderProgram.getValueId("a_Tangent"));
-    glEnableVertexAttribArray(shaderProgram.getValueId("a_Bitangent"));
+	//glUniform1f(ALGINE_CS_SWITCH_NORMAL_MAPPING, 1); // with mapping
+    glEnableVertexAttribArray(ALGINE_CS_IN_TANGENT);
+    glEnableVertexAttribArray(ALGINE_CS_IN_BITANGENT);
 	
-    for (size_t i = 0; i < MODELS_COUNT; i++) drawModel(shaderProgram, models[i]);
-	for (size_t i = 0; i < LAMPS_COUNT; i++) drawModel(shaderProgram, lamps[i]);
+    for (size_t i = 0; i < MODELS_COUNT; i++) drawModel(models[i]);
+	for (size_t i = 0; i < LAMPS_COUNT; i++) drawModel(lamps[i]);
 	
-    glDisableVertexAttribArray(shaderProgram.getValueId("a_Position"));
-    glDisableVertexAttribArray(shaderProgram.getValueId("a_Normal"));
-    glDisableVertexAttribArray(shaderProgram.getValueId("a_Texture"));
-    glDisableVertexAttribArray(shaderProgram.getValueId("a_Tangent"));
-    glDisableVertexAttribArray(shaderProgram.getValueId("a_Bitangent"));
+    glDisableVertexAttribArray(ALGINE_CS_IN_POSITION);
+    glDisableVertexAttribArray(ALGINE_CS_IN_NORMAL);
+    glDisableVertexAttribArray(ALGINE_CS_IN_TEXCOORD);
+    glDisableVertexAttribArray(ALGINE_CS_IN_TANGENT);
+    glDisableVertexAttribArray(ALGINE_CS_IN_BITANGENT);
 
-	advren::end();
+    end;
 }
+
+#undef begin
+#undef end
 
 void display() {
     /* --- shadow rendering --- */
-    glUseProgram(shaderShadowProgram.programId);
-	glUniform1f(shaderShadowProgram.getValueId("far_plane"), LAMP_FAR_PLANE);
-	for (GLuint i = 0; i < LAMPS_COUNT; i++) renderToDepthMap(shaderShadowProgram, i);
+    glUseProgram(ALGINE_PROGRAM_ID_SS);
+	glUniform1f(ALGINE_SS_FAR_PLANE, LAMP_FAR_PLANE);
+	for (GLuint i = 0; i < LAMPS_COUNT; i++) renderToDepthMap(i);
 	glUseProgram(0);
 		
 	/* --- color rendering --- */
-	glUseProgram(shaderProgram.programId);
+	glUseProgram(ALGINE_PROGRAM_ID_CS);
 	glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-	render(shaderProgram);
+	render();
 	glUseProgram(0);
 }
 
@@ -533,19 +579,30 @@ int main() {
     initShaders();
     initVAO();
     initMatrices();
-    advren::init();
-    advren::SCR_W = WIN_W;
-    advren::SCR_H = WIN_H;
+    ALGINE_SCR_W = WIN_W;
+    ALGINE_SCR_H = WIN_H;
     init_ModelBuffers_ModelMapping();
     createModels();
     initLamps();
     initShadowMaps();
-    mapping.loadLocations(shaderProgram.programId);
+    initShadowCalculation();
     initDOF();
 
     std::thread animate_scene_th(animate_scene);
 
+    double currentTime, previousTime = glfwGetTime();
+    size_t frameCount = 0;
     while (!glfwWindowShouldClose(window)) {
+        currentTime = glfwGetTime();
+        frameCount++;
+        // If a second has passed.
+        if (currentTime - previousTime >= 1.0) {
+            // Display the frame count here any way you want.
+            std::cout << frameCount << " fps\n";
+            frameCount = 0;
+            previousTime = currentTime;
+        }
+
         // Check if any events have been activiated (key pressed, mouse moved etc.) and call corresponding response functions
         glfwPollEvents();
         display();
@@ -566,20 +623,20 @@ int main() {
 void key_callback(GLFWwindow* window, int key, int scancode, int action, int mode) {
     if (key == GLFW_KEY_ESCAPE && action == GLFW_PRESS) glfwSetWindowShouldClose(window, GL_TRUE);
     else if ((key == GLFW_KEY_Q || key == GLFW_KEY_A || key == GLFW_KEY_W || key == GLFW_KEY_S) && action == GLFW_PRESS) {
-        glUseProgram(shaderProgram.programId);
+        glUseProgram(ALGINE_PROGRAM_ID_CS);
 
         if (key == GLFW_KEY_Q) {
             focalDepth++;
-            glUniform1f(shaderProgram.getValueId("focalDepth"), focalDepth);
+            glUniform1f(ALGINE_CS_FOCAL_DEPTH, focalDepth);
         } else if (key == GLFW_KEY_A) {
             focalDepth--;
-            glUniform1f(shaderProgram.getValueId("focalDepth"), focalDepth);
+            glUniform1f(ALGINE_CS_FOCAL_DEPTH, focalDepth);
         } else if (key == GLFW_KEY_W) {
             focalRange += 5;
-            glUniform1f(shaderProgram.getValueId("focalRange"), focalRange);
+            glUniform1f(ALGINE_CS_FOCAL_RANGE, focalRange);
         } else if (key == GLFW_KEY_S) {
             focalRange -= 5;
-            glUniform1f(shaderProgram.getValueId("focalRange"), focalRange);
+            glUniform1f(ALGINE_CS_FOCAL_RANGE, focalRange);
         }
 
         glUseProgram(0);